// packages/codecs/src/swf/types.ts
/**
 * SWF (Small Web Format / Shockwave Flash) format types and constants
 */

// SWF signatures
export const SWF_UNCOMPRESSED = 'FWS' // Flash Web Service
export const SWF_ZLIB = 'CWS' // Compressed with zlib
export const SWF_LZMA = 'ZWS' // Compressed with LZMA

// SWF version constants
export const MIN_SWF_VERSION = 1
export const MAX_SWF_VERSION = 40

// Tag types
export enum SwfTagType {
	End = 0,
	ShowFrame = 1,
	DefineShape = 2,
	PlaceObject = 4,
	RemoveObject = 5,
	DefineBits = 6,
	DefineButton = 7,
	JPEGTables = 8,
	SetBackgroundColor = 9,
	DefineFont = 10,
	DefineText = 11,
	DoAction = 12,
	DefineFontInfo = 13,
	DefineSound = 14,
	StartSound = 15,
	DefineButtonSound = 17,
	SoundStreamHead = 18,
	SoundStreamBlock = 19,
	DefineBitsLossless = 20,
	DefineBitsJPEG2 = 21,
	DefineShape2 = 22,
	DefineButtonCxform = 23,
	Protect = 24,
	PlaceObject2 = 26,
	RemoveObject2 = 28,
	DefineShape3 = 32,
	DefineText2 = 33,
	DefineButton2 = 34,
	DefineBitsJPEG3 = 35,
	DefineBitsLossless2 = 36,
	DefineSprite = 39,
	FrameLabel = 43,
	SoundStreamHead2 = 45,
	DefineMorphShape = 46,
	DefineFont2 = 48,
	ExportAssets = 56,
	ImportAssets = 57,
	EnableDebugger = 58,
	DoInitAction = 59,
	DefineVideoStream = 60,
	VideoFrame = 61,
	DefineFontInfo2 = 62,
	EnableDebugger2 = 64,
	ScriptLimits = 65,
	SetTabIndex = 66,
	FileAttributes = 69,
	PlaceObject3 = 70,
	ImportAssets2 = 71,
	DefineFontAlignZones = 73,
	CSMTextSettings = 74,
	DefineFont3 = 75,
	SymbolClass = 76,
	Metadata = 77,
	DefineScalingGrid = 78,
	DoABC = 82,
	DefineShape4 = 83,
	DefineMorphShape2 = 84,
	DefineSceneAndFrameLabelData = 86,
	DefineBinaryData = 87,
	DefineFontName = 88,
	StartSound2 = 89,
	DefineBitsJPEG4 = 90,
	DefineFont4 = 91,
}

/**
 * SWF file header
 */
export interface SwfHeader {
	signature: string // 'FWS', 'CWS', or 'ZWS'
	version: number
	fileLength: number
	frameSize: SwfRect
	frameRate: number
	frameCount: number
}

/**
 * SWF rectangle (in TWIPS - 1/20th of a pixel)
 */
export interface SwfRect {
	xMin: number
	xMax: number
	yMin: number
	yMax: number
}

/**
 * SWF RGB color
 */
export interface SwfRGB {
	red: number
	green: number
	blue: number
}

/**
 * SWF RGBA color
 */
export interface SwfRGBA extends SwfRGB {
	alpha: number
}

/**
 * SWF tag structure
 */
export interface SwfTag {
	type: SwfTagType
	data: Uint8Array
}

/**
 * Background color tag
 */
export interface SetBackgroundColorTag {
	type: SwfTagType.SetBackgroundColor
	color: SwfRGB
}

/**
 * Show frame tag
 */
export interface ShowFrameTag {
	type: SwfTagType.ShowFrame
}

/**
 * Define bits lossless tag (bitmap)
 */
export interface DefineBitsLosslessTag {
	type: SwfTagType.DefineBitsLossless | SwfTagType.DefineBitsLossless2
	characterId: number
	bitmapFormat: number
	bitmapWidth: number
	bitmapHeight: number
	bitmapData: Uint8Array
}

/**
 * Place object tag
 */
export interface PlaceObjectTag {
	type: SwfTagType.PlaceObject | SwfTagType.PlaceObject2 | SwfTagType.PlaceObject3
	depth: number
	characterId?: number
	matrix?: SwfMatrix
	colorTransform?: SwfColorTransform
}

/**
 * SWF matrix for transformations
 */
export interface SwfMatrix {
	scaleX: number
	scaleY: number
	rotateSkew0: number
	rotateSkew1: number
	translateX: number
	translateY: number
}

/**
 * SWF color transformation
 */
export interface SwfColorTransform {
	redMultiplier: number
	greenMultiplier: number
	blueMultiplier: number
	alphaMultiplier: number
	redAddition: number
	greenAddition: number
	blueAddition: number
	alphaAddition: number
}

/**
 * Complete SWF structure
 */
export interface SwfFile {
	header: SwfHeader
	tags: SwfTag[]
	backgroundColor?: SwfRGB
}

/**
 * TWIPS conversion constant (1/20th of a pixel)
 */
export const TWIPS_PER_PIXEL = 20